+++ scripts/createDmg.sh
@@ -6,8 +6,9 @@
 # Created by Derek Clarkson on 27/08/10.
 # Copyright 2010 Derek Clarkson. All rights reserved.
 
-echo "Copying markdown documentation from root of project ..."
+echo "Copying other documentation from root of project ..."
 cp -v *.markdown "$DC_ARTIFACT_DIR"
+cp -v *.textile "$DC_ARTIFACT_DIR"
 
 echo "Copying other files copied by build ..."
 find "$DC_BUILD_DIR/$DC_BUILD_CONFIGURATION-iphoneos" -type f -not -name "*.a" -depth 1 -exec cp -v "{}" "$DC_ARTIFACT_DIR" \;