+++ src/tests/RatingTestAppDelegate.m
@@ -15,7 +15,7 @@
 @synthesize ratingController;
 
 - (void) applicationDidFinishLaunching:(UIApplication *)application {
-	[window addSubview:ratingController.view];
+	//[window addSubview:ratingController.view];
 	[window makeKeyAndVisible];
 	DC_LOG(@"Main window loaded");
 }